+++ src/index.js
@@ -1,40 +1,4 @@
 import './style.css';
+import { loadImage } from './modules/Api';
 
-let pathImage;
-async function fetchApi() {
-  const response = await fetch('https://api.tvmaze.com/search/shows?q=girls');
-  const data = await response.json();
-  console.log(data);
-  return data;
-}
-
-async function loadImage() {
-  const arr = await fetchApi();
-  for (let i = 0; i < arr.length; i += 1) {
-    const divConatinerSingleImage = document.createElement('div');
-    pathImage = arr[i].show.image.medium;
-    const img = document.createElement('img');
-    const ContainerAllImages = document.querySelector('.hero');
-    img.src = pathImage;
-    divConatinerSingleImage.appendChild(img);
-    ContainerAllImages.appendChild(divConatinerSingleImage);
-    const containerLike = document.createElement('div');
-    containerLike.classList.add('containerLike');
-    const nameShow = arr[i].show.name;
-    const nameShowparagraph = document.createElement('p');
-    nameShowparagraph.textContent = nameShow;
-    containerLike.appendChild(nameShowparagraph);
-    divConatinerSingleImage.appendChild(containerLike);
-    const containerHearth = document.createElement('div');
-    containerHearth.classList.add('containerHearth');
-    const Hearth = document.createElement('span');
-    Hearth.classList.add('hearth');
-    Hearth.innerHTML = "<i class='bx bxs-heart'></i>";
-    containerHearth.appendChild(Hearth);
-    containerLike.appendChild(containerHearth);
-    const commentButton = document.createElement('button');
-    commentButton.textContent = 'Comment';
-    containerLike.appendChild(commentButton);
-  }
-}
 loadImage();
